+++ tests/Fixer/PhpdocParamTypeFixerTest.php
@@ -33,7 +33,7 @@ final class PhpdocParamTypeFixerTest extends AbstractFixerTestCase
      *
      * @dataProvider provideFixCases
      */
-    public function testFix(string $expected, string $input = null): void
+    public function testFix(string $expected, ?string $input = null): void
     {
         $this->doTest($expected, $input);
     }