+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## [Unreleased]
 - Add NoCommentedOutCodeFixer
+- Add NullableParamStyleFixer
+- Fix: NoImportFromGlobalNamespaceFixer - class without  namespace
+- Fix: NoUselessClassCommentFixer - comment detection
+- Fix: TokenRemover - remove last element of file
+- Fix: TokenRemover - remove item in line after code
+- Fix: NoImportFromGlobalNamespaceFixer - constant named the same as global imported class
 
 ## v1.11.0 - *2018-10-14*
 - Add PhpdocParamOrderFixer